Update bsp package to use enhanced wget command

+++ src/bsp/imx_firmware.mk
@@ -19,9 +19,12 @@ imx_firmware:
 	 \
 	 echo Installing firmware-imx for ddr,hdmi,dp,vpu,easrc,epdc,xcvr,xuvi && \
 	 if [ ! -d $(BSPDIR)/firmware-imx ]; then \
-	     cd $(BSPDIR) && wget -q $(repo_firmware_imx_bin_url) -O firmware_imx.bin $(LOG_MUTE) && \
-	     chmod +x firmware_imx.bin && \
-	     ./firmware_imx.bin --auto-accept $(LOG_MUTE) && mv firmware-imx* firmware-imx && rm -f firmware_imx.bin; \
+	     cd $(BSPDIR) && rm -f firmware_imx.bin; \
+		 $(WGET) $(repo_firmware_imx_bin_url) -O firmware_imx.bin $(LOG_MUTE); \
+		 [ $$? -ne 0 ] && { echo "Downloading $(repo_firmware_imx_bin_url) failed."; exit 1; } || \
+	     chmod +x firmware_imx.bin; \
+	     ./firmware_imx.bin --auto-accept --force $(LOG_MUTE); \
+		 mv firmware-imx* firmware-imx && rm -f firmware_imx.bin; \
 	 fi && \
 	 cp -Prf $(BSPDIR)/firmware-imx/firmware/* $(FBOUTDIR)/bsp/imx_firmware/lib/firmware/imx/ && \
 	 $(call fbprint_d,"imx_firmware")
